# packages/backend-core/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/orgs/me/members": {
      "get": {
        "operationId": "MembersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Members"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/orgs/me/members/{userId}": {
      "patch": {
        "operationId": "MembersController_patch",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMemberBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Members"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "delete": {
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Members"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
    "/v1/outreach/proposals": {
      "get": {
        "operationId": "OutreachProposalsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/outreach/proposals/{id}": {
      "get": {
        "operationId": "OutreachProposalsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "patch": {
        "operationId": "OutreachProposalsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProposalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/outreach/proposals/{id}/approve": {
      "post": {
        "operationId": "OutreachProposalsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/outreach/proposals/{id}/dismiss": {
      "post": {
        "operationId": "OutreachProposalsController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissProposalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
    "/v1/kb/curation/candidates": {
      "get": {
        "operationId": "KbCandidatesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KbCandidates"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/kb/curation/candidates/{id}": {
      "get": {
        "operationId": "KbCandidatesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KbCandidates"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "patch": {
        "operationId": "KbCandidatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KbCandidates"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/kb/curation/candidates/{id}/publish": {
      "post": {
        "operationId": "KbCandidatesController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishCandidateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KbCandidates"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/kb/curation/candidates/{id}/dismiss": {
      "post": {
        "operationId": "KbCandidatesController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KbCandidates"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/kb/spaces": {
      "get": {
        "operationId": "KbSpacesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KbSpaces"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Munin",
    "description": "HTTP surface area for the Munin platform. Authenticate with a bearer token (admin API key or delegated end-user token) or a session cookie.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "munin_session"
      }
    },
    "schemas": {
...
      "PatchMemberBody": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
...
      "UpdateProposalBody": {
        "type": "object",
        "properties": {
          "draftSubject": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "draftBody": {
            "type": "string",
            "minLength": 1,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        }
      },
      "DismissProposalBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
      "UpdateCandidateBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "body": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "PublishCandidateBody": {
        "type": "object",
        "properties": {
          "targetSpaceSlug": {
            "type": "string",
            "minLength": 1
          },
          "audiences": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "targetSpaceSlug"
        ]
      }
    }
  }
}